backend/requirements.txt
celery~=5.2.3
docx==0.2.4
Flask==2.1.2
Flask_Cors==3.0.10
jiagu==0.2.3
jieba==0.42.1
PyJWT==2.4.0
Werkzeug==2.1.2

psutil~=5.9.1
PyMySQL~=1.0.2
requests~=2.27.1